fix(kimi): drop unnecessary usize cast in load_messages round-trip test
CI runs `cargo clippy --all-targets --all-features -- -D warnings`,
which flags the local `len() as usize` no-op. The local check used
`cargo clippy --all-targets` without `-D warnings` and silently allowed it.

+++ src-tauri/tests/parser_tests.rs
@@ -984,7 +984,7 @@ fn kimi_load_messages_round_trips_through_source_path() {
     let loaded = provider
         .load_messages(&s.meta.id, &s.meta.source_path)
         .expect("load_messages");
-    assert_eq!(loaded.messages.len(), s.messages.len() as usize);
+    assert_eq!(loaded.messages.len(), s.messages.len());
 }
 
 #[test]
